Split the unique test file into others
This is a possible division of the responsibilities:
 - hooks: set up and tear down
 - commands: custom client actions
 - user_data: object to access and remove user data easily
